+++ infra/resources/prod/rbac.tf
@@ -5,6 +5,21 @@ module "roles" {
   principal_id    = module.backend_func.cdc_backend_func.principal_id
   subscription_id = data.azurerm_subscription.current.subscription_id
 
+  storage_blob = [
+    {
+      storage_account_name = module.storage_audit.immutable_cdc_audit_logs_storage.name
+      resource_group_name  = module.storage_audit.immutable_cdc_audit_logs_storage.resource_group_name
+      role                 = "owner"
+      description          = "Allow backend func to write FIMS audit blobs"
+    },
+    {
+      storage_account_name = module.storage_audit_proxy.immutable_cdc_audit_logs_storage_proxy.name
+      resource_group_name  = module.storage_audit_proxy.immutable_cdc_audit_logs_storage_proxy.resource_group_name
+      role                 = "owner"
+      description          = "Allow backend func to write external audit blobs"
+    }
+  ]
+
   storage_queue = [
     {
       storage_account_name = module.storage_be.cdc_storage_be.name
@@ -36,6 +51,13 @@ module "roles" {
       }
     }
   ]
+
+  cosmos = [{
+    account_name        = module.cosmos_db.cosmos_db.name
+    resource_group_name = module.cosmos_db.cosmos_db.resource_group_name
+    role                = "writer"
+    description         = "Allow CDC Backend Function to read/write CosmosDB via RBAC"
+  }]
 }
 
 module "roles_staging" {
@@ -44,6 +66,21 @@ module "roles_staging" {
 
   principal_id    = module.backend_func.cdc_backend_func.staging_principal_id
   subscription_id = data.azurerm_subscription.current.subscription_id
+
+  storage_blob = [
+    {
+      storage_account_name = module.storage_audit.immutable_cdc_audit_logs_storage.name
+      resource_group_name  = module.storage_audit.immutable_cdc_audit_logs_storage.resource_group_name
+      role                 = "owner"
+      description          = "Allow backend func (staging) to write FIMS audit blobs with index tags"
+    },
+    {
+      storage_account_name = module.storage_audit_proxy.immutable_cdc_audit_logs_storage_proxy.name
+      resource_group_name  = module.storage_audit_proxy.immutable_cdc_audit_logs_storage_proxy.resource_group_name
+      role                 = "owner"
+      description          = "Allow backend func (staging) to write external audit blobs with index tags"
+    }
+  ]
 
   storage_queue = [
     {
@@ -76,6 +113,14 @@ module "roles_staging" {
       }
     }
   ]
+
+  cosmos = [{
+    account_name        = module.cosmos_db.cosmos_db.name
+    resource_group_name = module.cosmos_db.cosmos_db.resource_group_name
+    role                = "writer"
+    description         = "Allow CDC Backend Function (staging) to read/write CosmosDB via RBAC"
+  }]
+
 }
 
 module "support_func_roles" {
@@ -95,6 +140,13 @@ module "support_func_roles" {
       }
     }
   ]
+
+  cosmos = [{
+    account_name        = module.cosmos_db.cosmos_db.name
+    resource_group_name = module.cosmos_db.cosmos_db.resource_group_name
+    role                = "writer"
+    description         = "Allow CDC Support Function to read/write CosmosDB via RBAC"
+  }]
 }
 
 module "support_func_staging_roles" {
@@ -114,4 +166,11 @@ module "support_func_staging_roles" {
       }
     }
   ]
+
+  cosmos = [{
+    account_name        = module.cosmos_db.cosmos_db.name
+    resource_group_name = module.cosmos_db.cosmos_db.resource_group_name
+    role                = "writer"
+    description         = "Allow CDC Support Function (staging) to read/write CosmosDB via RBAC"
+  }]
 }